update(console): separate propmpts and command tests
